requirements.txt
torch>=2.0.0
torchvision>=0.16.0
transformers>=4.0.0
diffusers>=0.24.0
compel>=2.0.0
tqdm>=4.0.0
Pillow>=10.0.0
numpy>=1.22.0
matplotlib>=3.6.0
opencv-python>=4.7.0
accelerate>=0.25.0
scipy>=1.10.0

# dev
pre-commit>=3.2.1
toml>=0.10.2
